+++ src/styles/RestaurantDetail.css
@@ -113,3 +113,14 @@ th {
 .menu-item:hover {
   transform: scale(1.05);
 }
+
+.btn-update{
+  background-color: #ff9800;
+  color: white;
+  border: none;
+  padding: 5px 10px;
+  font-size: 10px;
+  border-radius: 4px;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
